+++ Models/Discount_Detail.Model.cs
@@ -10,6 +10,6 @@ namespace uni_cap_pro_be.Models
 		[Required]
 		public required Guid DiscountId { get; set; }
 		public int Level { get; set; }
-		public double Amount { get; set; }
+		public float Amount { get; set; }
 	}
 }
